Format cleanup. - fixed ports.c to use a more portable method of getting integral pthread_t representation; resolves mingw crash. - added GCC attribute to generate compiler warnings for invalid usage of hb_log, hb_deep_log and hb_errror;   see new macro HB_WPRINTF(s,v) in common.h. - fixed various invalid usage of above functions on osx i386/x86_64, and mingw.
git-svn-id: svn://svn.handbrake.fr/HandBrake/trunk@2618 b64f7644-9d1e-0410-96f1-a4d463321fa5

+++ libhb/decsrtsub.c
@@ -123,8 +123,7 @@ static hb_buffer_t *srt_read( hb_work_private_t *pv )
                 hb_error( "Invalid shift sequence" );
             } else if ( ( retval == -1 ) && ( errno == EILSEQ ) )
             {
-                hb_error( "Invalid byte for codeset in input, %d bytes discarded",
-                          in_size);
+                hb_error( "Invalid byte for codeset in input, %"PRId64" bytes discarded", (int64_t)in_size);
             } else if ( ( retval == -1 ) && ( errno == E2BIG ) )
             {
                 hb_error( "Not enough space in output buffer");
